renamed navbar as header
