+++ backend/backend/src/index.js
@@ -1,5 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./components/Navbar/App";
+import App from "./components/Header/App";
 
 ReactDOM.render(<App />, document.getElementById("root"));
